backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
python-multipart==0.0.6
Pillow==10.1.0
opencv-python-headless==4.8.1.78
numpy==1.26.2
torch==2.1.1
torchvision==0.16.1
ultralytics==8.0.206
geopandas==0.14.1
folium==0.15.0
shapely==2.0.2
geopy==2.4.1
httpx==0.25.2
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
aiofiles==23.2.1
boto3==1.33.6
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
faker==20.1.0
